[CELEBORN-2015] Retry IOException failures for RPC requests
### What changes were proposed in this pull request?
- Support retry on IOException failures for RpcRequest in addition with RpcTimeoutException.
- Moved duplicate code to Utils

### Why are the changes needed?

Currently if a request fails with SocketException or IOException it does not get retried which leads to stage failures. Celeborn should retry on such connection failures.

### Does this PR introduce _any_ user-facing change?
NA

### How was this patch tested?
NA

Closes #3286 from s0nskar/setup_lifecycle_exception.

Authored-by: Sanskar Modi <sanskarmodi97@gmail.com>
(cherry picked from commit 612464c69deba27047d725dbcd25cdb8f8b994e5)
